+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 - Replace QUAST with MetaQUAST
 - Add species ID to run summary table
+- `-profile conda` is no longer supported, users should use `-profile standard` (Docker) or `-profile singularity` instead
+### Added
+- `nextflow run epi2me-labs/wf-human-variation --version` will now print the workflow version number and exit
 ### Fixes
 - Prokka only runs in denovo assembly mode
 - Tidy up report code
+  
 ## [v0.2.6]
 ### Changes
 - Added QUAST for assembly stats